LayoutTests/imported/w3c/web-platform-tests/svg/animations/svgpath-animation-1-expected.txt


FAIL Test 'by' animation on path. You should see a green 100x100 path and only PASS messages assert_equals: expected "M 40 40 L 109.975 40 L 109.975 109.975 L 40 109.975 Z" but got "M 40 40 L 60 40 L 60 60 L 40 60 z"
